JAVA 예제 업데이트 ( Calculator Update )

+++ src/Calculator.java
@@ -5,7 +5,7 @@ public class Calculator {
     double a;
     double b;
 
-    public Calculator  ( int a, int b ) {
+    public Calculator  ( double a, double b ) {
         this.a = a;
         this.b = b;
     }
